refactor: 에러 테스트를 어노테이션 기반에서 패키지 기반으로 변경
- 테스트용 어노테이션을 만들었던 것을 패키지로 나누어서 리플렉션을 사용하도록 변경
