+++ nextstep-di/src/test/java/nextstep/di/factory/BeanFactoryErrorTest.java
@@ -1,10 +1,7 @@
 package nextstep.di.factory;
 
-import nextstep.di.factory.error.BadBean;
-import nextstep.di.factory.error.Circular;
-import nextstep.di.factory.error.GoodBean;
-import nextstep.di.factory.error.OneConstructorBean;
-import org.junit.jupiter.api.BeforeEach;
+import nextstep.di.factory.beans.noerror.OneConstructorBean;
+import nextstep.stereotype.Controller;
 import org.junit.jupiter.api.Test;
 import org.reflections.Reflections;
 
@@ -17,14 +14,10 @@ public class BeanFactoryErrorTest {
     private BeanFactory beanFactory;
     private Reflections reflections;
 
-    @BeforeEach
-    void setUp() {
-        reflections = new Reflections("nextstep.di.factory.error");
-    }
-
     @Test
     void 기본_생성자만_있는_클래스_빈_생성_확인() {
-        Set<Class<?>> types = reflections.getTypesAnnotatedWith(GoodBean.class);
+        reflections = new Reflections("nextstep.di.factory.beans.noerror");
+        Set<Class<?>> types = reflections.getTypesAnnotatedWith(Controller.class);
         beanFactory = new BeanFactory(types);
         beanFactory.initialize();
         assertNotNull(beanFactory.getBean(OneConstructorBean.class));
@@ -32,14 +25,16 @@ public class BeanFactoryErrorTest {
 
     @Test
     void 순환참조_에러_확인() {
-        Set<Class<?>> types = reflections.getTypesAnnotatedWith(Circular.class);
+        reflections = new Reflections("nextstep.di.factory.beans.circular");
+        Set<Class<?>> types = reflections.getTypesAnnotatedWith(Controller.class);
         beanFactory = new BeanFactory(types);
         assertThatThrownBy(() -> beanFactory.initialize()).isInstanceOf(BeanCreateException.class);
     }
 
     @Test
     void 여러_생성자_중_결정할_수_없는_경우() {
-        Set<Class<?>> types = reflections.getTypesAnnotatedWith(BadBean.class);
+        reflections = new Reflections("nextstep.di.factory.beans.error");
+        Set<Class<?>> types = reflections.getTypesAnnotatedWith(Controller.class);
         beanFactory = new BeanFactory(types);
         assertThatThrownBy(() -> beanFactory.initialize()).isInstanceOf(BeanCreateException.class);
     }
