+++ nextstep-di/src/test/java/nextstep/di/factory/beans/circular/CircularReferenceBean3.java
@@ -1,6 +1,8 @@
-package nextstep.di.factory.error;
+package nextstep.di.factory.beans.circular;
 
-@Circular
+import nextstep.stereotype.Controller;
+
+@Controller
 public class CircularReferenceBean3 {
     private CircularReferenceBean circularReferenceBean;
