+++ nextstep-di/src/test/java/nextstep/di/factory/beans/noerror/ExampleBean.java
@@ -1,0 +1,7 @@
+package nextstep.di.factory.beans.noerror;
+
+import nextstep.stereotype.Controller;
+
+@Controller
+public class ExampleBean {
+}
