--- nextstep-di/src/test/java/nextstep/di/factory/error/BadBean.java
@@ -1,4 +1,0 @@
-package nextstep.di.factory.error;
-
-public @interface BadBean {
-}
