--- nextstep-di/src/test/java/nextstep/di/factory/error/ExampleBean.java
@@ -1,5 +1,0 @@
-package nextstep.di.factory.error;
-
-@GoodBean
-public class ExampleBean {
-}
